create popover base for edit and delete popovers
